refactor: toLog to class fix: 500 errors
