+++ src/main/java/ru/practicum/shareit/aspect/LoggingAspect.java
@@ -1,17 +1,19 @@
-package ru.practicum.shareit.logging;
+package ru.practicum.shareit.aspect;
 
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
+import org.springframework.stereotype.Component;
 
 import java.util.Arrays;
 
 @Aspect
+@Component
 @Slf4j
 public class LoggingAspect {
 
-    @Around("@annotation(ToLog)")
+    @Around("@within(ru.practicum.shareit.marker.ToLog)")
     public Object log(ProceedingJoinPoint joinPoint) throws Throwable {
         String methodName = joinPoint.getSignature().getName();
         Object[] arguments = joinPoint.getArgs();
