+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import ru.practicum.shareit.booking.dto.CreateBookingDto;
 import ru.practicum.shareit.booking.dto.GetBookingDto;
-import ru.practicum.shareit.logging.ToLog;
+import ru.practicum.shareit.marker.ToLog;
 import ru.practicum.shareit.validator.ValuesAllowedConstraint;
 
 import javax.validation.Valid;
@@ -25,11 +25,11 @@ import static ru.practicum.shareit.util.Constants.REQUEST_HEADER_USER_ID;
 @RequestMapping(path = "/bookings")
 @RequiredArgsConstructor
 @Validated
+@ToLog
 public class BookingController {
     private final BookingService bookingService;
 
     @GetMapping
-    @ToLog
     public List<GetBookingDto> getUserBookings(@RequestHeader(REQUEST_HEADER_USER_ID) long userId,
                                                @ValuesAllowedConstraint(propName = "state",
                                                        values = {"all",
@@ -44,7 +44,6 @@ public class BookingController {
     }
 
     @GetMapping("/owner")
-    @ToLog
     public List<GetBookingDto> getOwnerBookings(@RequestHeader(REQUEST_HEADER_USER_ID) long userId,
                                                 @ValuesAllowedConstraint(propName = "state",
                                                         values = {"all",
@@ -59,21 +58,18 @@ public class BookingController {
     }
 
     @GetMapping("/{bookingId}")
-    @ToLog
     public GetBookingDto getBookingByUserOwner(@RequestHeader(REQUEST_HEADER_USER_ID) long userId,
                                                @PathVariable long bookingId) {
         return bookingService.getBookingByUserOwner(userId, bookingId);
     }
 
     @PostMapping
-    @ToLog
     public GetBookingDto create(@RequestHeader(REQUEST_HEADER_USER_ID) long userId,
                                 @RequestBody @Valid CreateBookingDto createBookingDto) {
         return bookingService.create(userId, createBookingDto);
     }
 
     @PatchMapping("/{bookingId}")
-    @ToLog
     public GetBookingDto approveBooking(@RequestHeader(REQUEST_HEADER_USER_ID) long userId,
                                         @PathVariable long bookingId,
                                         @RequestParam Boolean approved) {
