+++ src/main/java/ru/practicum/shareit/config/BeanConfig.java
@@ -4,7 +4,6 @@ import org.h2.tools.Server;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import ru.practicum.shareit.logging.LoggingAspect;
 
 import java.sql.SQLException;
 
@@ -15,10 +14,5 @@ public class BeanConfig {
     @Profile("test")
     public Server inMemoryH2DatabaseaServer() throws SQLException {
         return Server.createTcpServer("-tcp", "-tcpAllowOthers", "-tcpPort", "9090");
-    }
-
-    @Bean
-    public LoggingAspect loggingAspect() {
-        return new LoggingAspect();
     }
 }
