+++ src/main/java/ru/practicum/shareit/exception/ErrorHandler.java
@@ -3,6 +3,7 @@ package ru.practicum.shareit.exception;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.ServletRequestBindingException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -39,20 +40,6 @@ public class ErrorHandler {
         );
     }
 
-    @ExceptionHandler({ValidationException.class,
-            NotAvailableException.class,
-            NotValidDateException.class})
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public Map<String, String> handleValidation(final RuntimeException e) {
-        log.debug("Получен статус {} {}. Причина: {}",
-                HttpStatus.BAD_REQUEST.value(),
-                HttpStatus.BAD_REQUEST.getReasonPhrase(),
-                e.getMessage());
-        return Map.of(
-                "error", e.getMessage()
-        );
-    }
-
     @ExceptionHandler
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
     public Map<String, String> handleThrowable(final Throwable e) {
@@ -81,7 +68,11 @@ public class ErrorHandler {
     }
 
     @ExceptionHandler({MethodArgumentNotValidException.class,
-            MethodArgumentException.class})
+            MethodArgumentException.class,
+            ServletRequestBindingException.class,
+            ValidationException.class,
+            NotAvailableException.class,
+            NotValidDateException.class})
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public Map<String, String> handleThrowable(final RuntimeException e) {
         log.debug("Получен статус {} {}. Причина: {}",
