+++ src/main/java/ru/practicum/shareit/marker/ToLog.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.logging;
+package ru.practicum.shareit.marker;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,6 +6,6 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
+@Target({ElementType.TYPE, ElementType.METHOD})
 public @interface ToLog {
 }
